change in select updates state
